Major version up to v1.0.0
The purpose of the major version up is to cater both
synchronous and asynchronous transactions for curl-rust.

There are breaking changes in terms of the builder pattern.
We use blocking() to signify the end of the Curl builder to do a
blocking perform after and nonblocking() to signify the end of the
Curl builder to do a nonblocking perform after.

For nonblocking transaction, the actor handler is now passed into nonblocking
function instead in the new constructor of the HttpClient.